spa refactor WP2: app shell
Decompose js/base/base.js into js/app/* modules (spinner, materialize
compat, dialog, theme, triggers, i18n, bootstrap fetch/cache, nav builder,
router, and the three shared modals now populated from /xhr/bootstrap
instead of Jinja loops) plus a hand-rolled History-API router with the
view-lifecycle contract from PLAN_SPA_REFACTOR.md §2.2. Webpack collapses
to a single config with `app` + `public` entries (public.js is the small
i18n/theme bundle for login/auth_error/ical_action); base.html gains a
`bundle` block so each entry's headers/*.html loads in the same early
head position the old base bundle did (preserves the CSS cascade order
against theme.css). A trivial index view proves the router end-to-end,
including the client-side default-plan redirect once /plan/:pid exists.

Also fixes containers/Dockerfile(_debug), which only COPYed js/views and
js/base into the build stage — missing the new js/app and js/lib dirs
would have broken every container build from here on.

Verified against the debug e2e container: full login.spec.ts (8/8),
settings/{preferences,calendar,password,logout}.spec.ts (64/64) —
exercising the dropdown menus, all three shared modals end-to-end in a
real browser, not just render checks — and the 101 pytest tests, all
green. endpoints-coverage.spec.ts now additionally covers /xhr/bootstrap;
its remaining 4 new-endpoint entries are deferred to the work packages
that convert their views (WP5-7), per PLAN_SPA_REFACTOR.md §5.

Part of PLAN_SPA_REFACTOR.md.
