+++ js/app/main.js
@@ -1,0 +1,69 @@
+'use strict';
+
+// Boot sequence — see PLAN_SPA_REFACTOR.md §2.4.
+// publicPath must be the very first import: it must run before any other
+// module's side effects could trigger a dynamic chunk load.
+import './publicPath.js';
+import '../base/style.css';
+
+import loadI18n from './i18n.js';
+import * as bootstrap from './bootstrap.js';
+import { M } from './materialize.js';
+import './dialog.js';
+import { initThemeToggle } from './theme.js';
+import { initTriggerClasses } from './triggers.js';
+import { initPrefs } from './modals/prefs.js';
+import { initCalendar } from './modals/calendar.js';
+import { initChangePassword } from './modals/changePassword.js';
+import * as nav from './nav.js';
+import router from './router.js';
+
+// Materialize 2.x needs an explicit init call to wire up .dropdown-trigger
+// elements (the user/admin menu icons) — there's no auto-init class scan.
+function initDropdowns() {
+  document.querySelectorAll('.dropdown-trigger').forEach(function (el) {
+    M.Dropdown.init(el, { coverTrigger: false, constrainWidth: false });
+  });
+}
+
+async function boot() {
+  let pendingToast = window.sessionStorage.getItem('pendingToast');
+
+  // In parallel: i18n and the nav/prefs/plan bootstrap payload. Both are
+  // awaited below — the calendar modal reads window.warpGlobals.i18n
+  // synchronously at init, so TR must be ready before initCalendar().
+  await Promise.all([loadI18n(), bootstrap.get()]);
+
+  initDropdowns();
+  initThemeToggle();
+  initTriggerClasses();
+  initPrefs();
+  initCalendar();
+  initChangePassword();
+
+  window.TR.updateDOM();
+
+  // Opt every WARP form field into Materialize 2.x's built-in `.outlined`
+  // text-field variant (bordered box). Scoped to .warp-fields containers so
+  // the nav search and the zone "book-as" underline field are left untouched;
+  // chips are excluded (they are a multi-value container, not a text field).
+  document.querySelectorAll('.warp-fields .input-field:not(.chips)').forEach(function (el) {
+    el.classList.add('outlined');
+  });
+  M.updateTextFields(); // ensure placeholder=" " so 2.x CSS label-float works
+
+  await nav.render();
+
+  if (pendingToast) {
+    window.sessionStorage.removeItem('pendingToast');
+    M.toast({ text: pendingToast });
+  }
+
+  await router.start();
+}
+
+if (document.readyState === 'loading') {
+  document.addEventListener('DOMContentLoaded', boot);
+} else {
+  boot();
+}
