+++ js/views/html/index.html
@@ -1,0 +1,6 @@
+<div class="index_container">
+    <div>
+    <svg class="logo-hero" viewBox="0 0 390 85" role="img" aria-label="WARP"><use id="index-logo-use" href=""/></svg><br>
+    Workspace Autonomous Reservation Program
+    </div>
+</div>
